% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5C_subset_by_taxon.R
\name{flatten.to.single.level}
\alias{flatten.to.single.level}
\title{Flatten a "tax.list" to a "flat.list"}
\usage{
flatten.to.single.level(my.list, finest.level = length(my.list$bq))
}
\arguments{
\item{my.list}{A "tax.list" that has these names: "names", "av", "sd", "bq", "pd", "br", each of which is a nested list that has a matrix for each taxonomy level.}

\item{finest.level}{Either a numeric taxonomy level or a character taxon name. This defines the taxonomy level that the list is flattened to. By default, the OTU-level is chosen.}
}
\value{
A "flat.list" that includes these names: "names", "av", "sd", "bq", "pd", "br", each of which is an individual matrix at the chosen taxonomy level.
}
\description{
Take a nested list with every taxonomy level and flatten it into a single list with only 1 taxonomy level.
}
