% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5C_subset_by_taxon.R
\name{get.taxon.level}
\alias{get.taxon.level}
\title{Get the taxonomy level of a given taxon name}
\usage{
get.taxon.level(my.list, taxon)
}
\arguments{
\item{my.list}{Either a "tax.list" or a "flat.list"}

\item{taxon}{An exactly spelled taxon name. This is case sensitive.}
}
\value{
The numeric taxon level.
}
\description{
Get the numeric level of the provided taxon name. Kingdom = 1, Phylum = 2, Class = 3, Order = 4, Family/Lineage = 5, Genus/Clade = 6, Species/Tribe = 7, SeqID = 8.
}
