% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5B_subset_by_sample.R
\name{group.by.sample}
\alias{group.by.sample}
\title{Group by sample}
\usage{
group.by.sample(my.list, group.index.list, new.colname.vect)
}
\arguments{
\item{my.list}{Either a "tax.list" or a "flat.list."}

\item{group.index.list}{A list where each element is a vector of sample (column) indeces that should be grouped together.}

\item{new.colname.vect}{A vector of new column names. The length of this vector should match the length of the group.index.list.}
}
\value{
Either a "tax.list" or a "flat.list" where the the samples have been grouped together. Taxon abundances are summed, standard deviation is propagated, and quantification limits are recalculated.
}
\description{
Lump samples together into a single group. For example, you could group all June samples into one sample.
}
