% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5A_group_taxa.R
\name{group.by.tax}
\alias{group.by.tax}
\title{Group by taxonomy}
\usage{
group.by.tax(my.list, verbose = FALSE)
}
\arguments{
\item{my.list}{A "flat.list" that includes these names: "names", "av", "sd", "bq", "pd", "br", each of which is an individual matrix.}

\item{verbose}{Logical.}
}
\value{
A "tax.list" that has these names: "names", "av", "sd", "bq", "pd", "br", each of which is a nested list that has a matrix for each taxonomy level.
}
\description{
Group a "flat list" structure by taxonomy, so that it becomes a nested list with each taxon level included separately. In addition to summing daughter taxa abundances, this function propagates error for the sd and calculates quantification limits.
}
