% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{key}
\alias{key}
\title{key}
\format{
\subsection{\code{key}}{

\describe{
\item{in.R.colnames}{exact matches to the sample names in limony matrix columns}
\item{in.R.order}{the order of columns in the limony matrices}
\item{Year}{Sample year}
\item{Month}{Sample month}
\item{Day}{Sample day}
\item{Date}{Sample date, with time zone summer in Wisconsin}
\item{Invasion}{Invasion regime}
\item{Season}{Season as defined in Rohwer et al. 2023}
\item{DNA.ng.uL}{Concentration of DNA extraction (can be used to filter low-yield samples)}
\item{Rep.Type}{Type of replicate}
\item{Color.Invasion}{Colors used in Rohwer et al. 2023}
\item{Color.Season}{Colors used in Rohwer et al. 2023}
\item{yDay}{Julian day of year}
\item{Days.Since.Spring}{Days since the start of Spring}
\item{Days.Since.Clearwater}{Days since the start of Clearwater}
\item{Days.Since.Early.Summer}{Days since the start of Early Summer}
\item{Days.Since.Late.Summer}{Days since the start of Late Summer}
\item{Days.Since.Fall}{Days since the start of Fall}
\item{Days.Since.Ice.On}{Days since the start of the previous Ice-On}
}
}
}
\source{
Rohwer et al. Species invasions shift microbial phenology in a two-decade freshwater time series. bioRxiv. 2022
}
\usage{
key
}
\description{
A data frame that includes information about each sample
If you use the season information, please cite the paper that defined the seasons:
Rohwer et al. Species invasions shift microbial phenology in a two-decade freshwater time series. bioRxiv. 2022
}
\keyword{datasets}
