% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{limony}
\alias{limony}
\title{limony
The nested list structure that contains all OTU tables.
For a visual diagram, see Figure 3 in the limony data announcement.
Please cite the limony data announcement if you use this data:
Rohwer & McMahon. Lake iTag measurements over nineteen years, introducing the limony dataset. bioRxiv. 2022}
\format{
\subsection{\code{limony}}{

A list with 6 elements:
\describe{
\item{names}{Taxonomic Assignments (character matrices)}
\item{av}{average abundance (integer matrices)}
\item{sd}{standard deviation (integer matrices)}
\item{bq}{below quantification (boolean matrices)}
\item{pd}{partial detection (boolean matrices)}
\item{br}{below resolution (boolean matrices)}
}
}
}
\source{
Rohwer & McMahon. Lake iTag measurements over nineteen years, introducing the limony dataset. bioRxiv. 2022
}
\usage{
limony
}
\description{
limony
The nested list structure that contains all OTU tables.
For a visual diagram, see Figure 3 in the limony data announcement.
Please cite the limony data announcement if you use this data:
Rohwer & McMahon. Lake iTag measurements over nineteen years, introducing the limony dataset. bioRxiv. 2022
}
\keyword{datasets}
