% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5A_group_taxa.R
\name{make.empty.list.structure}
\alias{make.empty.list.structure}
\title{Make an empty list structure.}
\usage{
make.empty.list.structure(ListNames)
}
\arguments{
\item{ListNames}{A vector of names to use}
}
\value{
A list where each element is NULL but is named with the ListNames
}
\description{
Make a where each element is NULL but has the provided names.
}
