% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5C_subset_by_taxon.R
\name{make.zero.below.LOQ}
\alias{make.zero.below.LOQ}
\title{Make abundances below the limit of quantification equal zero}
\usage{
make.zero.below.LOQ(my.list, LOQ.def = "bq", renorm = TRUE, verbose = TRUE)
}
\arguments{
\item{my.list}{Either a "tax.list" or a "flat.list".}

\item{LOQ.def}{Either "bq", "br", or "pd"}

\item{renorm}{Logical. Should samples dates be renomralized to 100\% total abundance.}

\item{verbose}{Logical.}
}
\description{
If a limit of quantification definition is chosen, this will make all values below it equal to zero. The LOQ can be defined as below resolution (br): below to abundance of a single read in the shallowest sample, partially detected (pd): not detected in all replicates, or below quantification (bq): either br or pd.
}
