% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5C_subset_by_taxon.R
\name{print.names.under.taxon}
\alias{print.names.under.taxon}
\title{Print daughter taxon names}
\usage{
\method{print}{names.under.taxon}(
  my.list,
  taxon,
  lower.lvl = "print minimum",
  upper.lvl = F,
  show.lvls = F,
  sort.by = "alphabet"
)
}
\arguments{
\item{my.list}{Either a "tax.list" or a "flat.list"}

\item{taxon}{A correctly spelled and case sensitive taxon name.}

\item{lower.lvl}{The numeric lowest taxon level names to print. By default, one level below the provided taxon name is printed.}

\item{upper.lvl}{The numeric upper taxon level names to print. By default no parent taxa are displayed.}

\item{show.lvls}{FALSE or numeric vector. If a vector taxon levels is provided, it will print the taxon at those levels.}

\item{sort.by}{How to order the daughter taxa. Either a function such as max or mean, or the character strings "presence" or "alphabet." By default results are shown alphabetically.}
}
\description{
For the provided taxon name, print the names of its constituent taxa. This can be used to step down a taxonomy and find the correct spellings of daughter taxa.
}
