% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5C_subset_by_taxon.R
\name{renormalize}
\alias{renormalize}
\title{Normalize abundances by day total or by taxon max values.}
\usage{
renormalize(my.list, by.day = TRUE, by.taxon.max = FALSE, verbose = TRUE)
}
\arguments{
\item{my.list}{Either a "tax.list" or a "flat.list"}

\item{by.day}{Logical. Normalize abundances by sample date- each column divided by its sum. Columns will sum to 100\%.}

\item{by.taxon.max}{Logical. Normalize abundances by taxon max value- each row divided by its max. Rows will have a max value of 100\%.}

\item{verbose}{Logical. Print messages about what's being done.}
}
\value{
Either a "tax.list" or a "flat.list" that has been normalized.
After subsetting taxa
}
\description{
Normalize abundances by day total or by taxon max values.
}
