% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{seasons}
\alias{seasons}
\title{seasons
A data frame that includes start dates for each season each year
If you use the season information, please cite the paper that defined the seasons:
Rohwer et al. Species invasions shift microbial phenology in a two-decade freshwater time series. bioRxiv. 2022}
\format{
\subsection{\code{seasons}}{

\describe{
\item{Year}{The observation year}
\item{Spring}{Start dates for the season Spring}
\item{Clearwater}{Start dates for the season Clearwater}
\item{Early.Summer}{Start dates for the season Early Summer}
\item{Late.Summer}{Start dates for the season Late Summer}
\item{Fall}{Start dates for the season Fall}
\item{Ice.On}{Start dates for the season Ice-on}
}
}
}
\source{
Rohwer et al. Species invasions shift microbial phenology in a two-decade freshwater time series. bioRxiv. 2022
}
\usage{
seasons
}
\description{
seasons
A data frame that includes start dates for each season each year
If you use the season information, please cite the paper that defined the seasons:
Rohwer et al. Species invasions shift microbial phenology in a two-decade freshwater time series. bioRxiv. 2022
}
\keyword{datasets}
