% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5C_subset_by_taxon.R
\name{sort.by.abundance}
\alias{sort.by.abundance}
\title{Sort by abundance}
\usage{
\method{sort}{by.abundance}(my.list, sort.by = sum)
}
\arguments{
\item{my.list}{A "flat.list" structure}

\item{sort.by}{Any function that will be applied to summarize the abundance of the rows. For example, this could be sum, mean, median. It could also be the character string "presence" to sort by presence/absence.}
}
\value{
a flat list with rows sorted by abundance instead of alphabetically.
}
\description{
Sort the OTU table rows by the overall abundances of the OTU. Can use your preferred summary function.
}
