% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5C_subset_by_taxon.R
\name{subset.by.taxa}
\alias{subset.by.taxa}
\title{Subset by taxa}
\usage{
\method{subset}{by.taxa}(my.list, keep.index, renormalize = F, verbose = T)
}
\arguments{
\item{my.list}{A "flat.list" that includes these names: "names", "av", "sd", "bq", "pd", "br", each of which is an individual matrix.}

\item{keep.index}{A numeric vector of row indexes to keep.}

\item{renormalize}{Logical. Should samples be re-normalized to that abundances on a given day sum to 100\%.}

\item{verbose}{Logical.}
}
\description{
Subset a flat.list to include only the provided row indexes. Indexes can be generated by the get.taxon.indexes function. After subsetting, a tax.list can be regenerated with the group.by.tax function.
}
