% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5B_subset_by_sample.R
\name{get.sample.indexes}
\alias{get.sample.indexes}
\title{Get sample indexes}
\usage{
get.sample.indexes(
  my.list,
  start.YY.MM.DD = "start",
  end.YY.MM.DD = "end",
  dates.are.season.range = FALSE,
  remove.prefiltered = F,
  remove.diff.depths = F,
  remove.diff.loc = F,
  only.prefiltered = F,
  only.diff.depths = F,
  only.diff.loc = F,
  my.key = NA,
  remove.low.dna = F,
  low.yield.ng.uL = 3,
  only.low.dna = F,
  selected.seasons = NA,
  selected.invasions = NA
)
}
\arguments{
\item{my.list}{either a "tax.list" or a "flat.list", so a list with these names: these names: "names", "av", "sd", "bq", "pd", "br", each of which contains either a list of taxonomy matrices or a single matrix.}

\item{start.YY.MM.DD}{character ymd-formatted date, or "start" to indicate use the earliest date}

\item{end.YY.MM.DD}{character ymd-formatted date, or "end" to indicate use the latest date}

\item{dates.are.season.range}{logical. If FALSE, dates used are the range between the start and the end date. If TRUE, dates used are the range each year between the start and end month and day. If TRUE, the year must still be included but can be set to anything.}

\item{remove.prefiltered}{logical. Remove the non-standard samples that received 10um pre-filtration.}

\item{remove.diff.depths}{logical. Remove the non-standard samples that were taken at depths different than a 12 m integrated sample.}

\item{remove.diff.loc}{logical. Remove the non-standard samples that were taken at a different location that the deep hole.}

\item{only.prefiltered}{logical. Keep only the pre-filtered samples.}

\item{only.diff.depths}{logical. Keep only the non-standard depths.}

\item{only.diff.loc}{logical. Keep only the non-standard locations.}

\item{my.key}{the "key" dataframe supplied with this package.}

\item{remove.low.dna}{logical. Remove samples that had low-yield DNA extractions.}

\item{low.yield.ng.uL}{numeric. The yield cutoff for what counts as low (default = 3 ng/uL).}

\item{only.low.dna}{logical. Keep only low-yield samples.}

\item{selected.seasons}{character vector. Keep only the listed seasons. Season options are "spring", "clearwater", "early summer", "late summer", "fall", "ice-on".}

\item{selected.invasions}{character vector. Keep only the listed invasion regimes. Invasion options are "none", "spiny", "zebra".}
}
\value{
A vector of indexes that can be used in the function subset.by.sample()
}
\description{
Get the indexes of samples to either keep or remove. Samples can be selected based on date or based on known method irregularities.
}
